Fixes an issue where creating a new Single Player Profile with only character name and difficulty crashed D2BS (and Diablo 2)
The example below randomly crashes in the context of D2Bot ran script. The idea is to randomize the difficulty the bot is ran at to regenerate the single player map.

The crash occurs in the call to Profile() with only 3 parameters. Seems the usage of `wcslen` when the parameter is empty is causing it to fail and crash. The fix replaces `wcslen` with `_countof` as suggested in several Stack Overflow threads. This would work if all fields (account, password) are filled, but for Single Player they're not and it's causing a crash.

```
function getRandomInt(max) {
  return Math.floor(Math.random() * Math.floor(max));
}

while (true) {
	if (getLocation() == 8 || getLocation() == 12) {
		const newProfile = Profile(
			1,
			"borc-three",
			getRandomInt(3)
		);
		newProfile.login();
	}
}
```