Make possible selecting a character that's not visible on the screen.
This fix adds the option to automatically scan the full character list before giving up and stopping.

This allows selecting a character that's not initially visible on the screen, e.g. character number 12.

The way it works is it checks the last on-screen character position and if there is a character there, it tries to scroll by clicking the Down arrow on the sidebar.

If a new character appears in the last place, it tries to scan for the searched character again. This repeats until there's no more character in the last slot or if the last character doesn't change after scrolling.